+++ versions/v19/projects/sdcorejs-angular/components/query-bar/src/components/chip-popover/chip-popover.component.ts
@@ -49,7 +49,8 @@ interface Range {
   styleUrl: './chip-popover.component.scss',
   standalone: true,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [SdIcon,
+  imports: [
+    SdIcon,
     CommonModule,
     FormsModule,
     MatMenuModule,