+++ versions/v19/projects/sdcorejs-angular/forms/date-range/src/date-range.component.ts
@@ -68,7 +68,8 @@ interface Daterange {
     }),
   ],
   standalone: true,
-  imports: [SdIcon,
+  imports: [
+    SdIcon,
     CommonModule,
     FormsModule,
     ReactiveFormsModule,