+++ versions/v20/projects/sdcorejs-angular/components/mini-editor/src/index.ts
@@ -1,2 +1,2 @@
-export * from './mini-editor.model';
+export type * from './mini-editor.model';
 export * from './mini-editor.component';