+++ versions/v20/projects/sdcorejs-angular/components/org-chart/index.ts
@@ -1,3 +1,3 @@
 export * from './src/org-chart.component';
 export * from './src/org-chart-item-def.directive';
-export * from './src/org-chart.model';
+export type * from './src/org-chart.model';