+++ versions/v20/projects/sdcorejs-angular/forms/chip-calendar/src/chip-calendar.component.ts
@@ -73,7 +73,8 @@ class SdChipCalendarErrorStateMatcher implements ErrorStateMatcher {
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
   host: { '[class.sd-has-label]': '!!label()', '[class.sd-viewed]': 'isViewed()' },
-  imports: [SdIcon,
+  imports: [
+    SdIcon,
     CommonModule,
     FormsModule,
     ReactiveFormsModule,